attached_assets/requirements_1759338673897.txt
pip>=24.0
setuptools>=70.0.0
wheel
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.4
gunicorn==23.0.0
sentence-transformers
faiss-cpu
numpy
torch
transformers
google-genai